# API_Config/api/swagger.yaml
basePath: /api/config
definitions:
  models.Alerts:
    properties:
      email:
        type: string
      targets:
        items:
          type: string
        type: array
    type: object
  models.CustomError:
    properties:
      code:
        default: 200
        type: integer
      message:
        type: string
    type: object
  models.Resources:
    properties:
      name:
        type: string
      uid:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: thomas.blosse.gilio@gmail.com
    name: Thomas Blosse
  description: Microservice pour la gestion des configurations
  termsOfService: http://swagger.io/terms/
  title: Config API
  version: "1.0"
paths:
  /alerts:
    get:
      description: Retrieves all the alerts in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Alerts'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.CustomError'
      summary: Get all alerts
      tags:
      - alerts
    post:
      consumes:
      - application/json
      description: Creates a new alert from the JSON payload and saves it in the database
      parameters:
      - description: Alert to create
        in: body
        name: alert
        required: true
        schema:
          $ref: '#/definitions/models.Alerts'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Alerts'
        "400":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/models.CustomError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.CustomError'
      summary: Create a new alert
      tags:
      - alerts
  /alerts/{id}:
    delete:
      description: Deletes an alert associated with the authenticated user
      responses:
        "200":
          description: Alert deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.CustomError'
      summary: Delete alert
      tags:
      - alerts
    get:
      description: Retrieves alerts associated with a specific resource
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Alerts'
            type: array
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.CustomError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.CustomError'
      summary: Get alerts by resource ID
      tags:
      - alerts
    put:
      consumes:
      - application/json
      description: Updates the targets of an existing alert for the authenticated
        user
      produces:
      - application/json
      responses:
        "200":
          description: Alert updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.CustomError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.CustomError'
      summary: Update alert
      tags:
      - alerts
  /resources:
    get:
      description: Retrieves all the resources
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Resources'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.CustomError'
      summary: Get all resources
      tags:
      - resources
    post:
      consumes:
      - application/json
      description: Creates a new resource
      parameters:
      - description: New resource
        in: body
        name: resource
        required: true
        schema:
          $ref: '#/definitions/models.Resources'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Resources'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.CustomError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.CustomError'
      summary: Create resource
      tags:
      - resources
  /resources/{id}:
    delete:
      description: Deletes a resource by its ID
      responses:
        "200":
          description: Resource deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.CustomError'
      summary: Delete resource
      tags:
      - resources
    get:
      description: Retrieves a resource given its ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Resources'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.CustomError'
      summary: Get resource by ID
      tags:
      - resources
swagger: "2.0"